Split the class to 2 different functionality
