+++ pom.xml
@@ -11,7 +11,7 @@
     <properties>
 
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>8</java.version>
+        <java.version>11</java.version>
 
         <!--     Plugins Version       -->
         <maven.compiler.version>3.8.1</maven.compiler.version>
